Move remaining Windows-related state tests to pytest
